Refactor Adapters into Packages
